move duplicate **Scale code into GenericScale class
