+++ generic.py
@@ -207,7 +207,48 @@ class GenericDownload(object):
         self.directory = path.join(self.par.project_directory, name)  
         if path.isdir(self.directory) == False:
             makedirs(self.directory) 
+
+class GenericScale:
+    
+    def __init__(self, sfile):
+        # read parameter file
+        self.sfile = sfile
+        par = ParameterIO(self.sfile)
+        self.intpdir = path.join(par.project_directory, 'interpolated')
+        self.scdir = self.makeOutDir(par)
+        self.list_name = par.station_list.split(path.extsep)[0]
+        
+        # get the station file
+        self.stations_csv = path.join(par.project_directory,
+                                      'par', par.station_list)
+        #read station points 
+        self.stations = StationListRead(self.stations_csv)  
+        
+        # read kernels
+        self.kernels = par.kernels
+        if not isinstance(self.kernels, list):
+            self.kernels = [self.kernels]
+    
+    def getOutNCF(self, par, src, scaleDir = 'scale'):
+        '''make out file name'''
+        
+        timestep = str(par.time_step) + 'h'
+        src = '_'.join(['scaled', src, timestep])
+        src = src + '.nc'
+        fname = path.join(self.scdir, src)
+        
+        return fname
+    
+    def makeOutDir(self, par):
+        '''make directory to hold outputs'''
+        
+        dirSC = path.join(par.project_directory, 'scaled')
+        
+        if not (path.isdir(dirSC)):
+            makedirs(dirSC)
             
+        return dirSC
+        
 def variables_skip(variable_name):
         '''
         Which variable names to use? Drop the ones that are dimensions.
