+++ merra_2.py
@@ -112,7 +112,7 @@ from os                import path, listdir, makedirs, remove
 from netCDF4           import Dataset, MFDataset
 from dateutil.rrule    import rrule, DAILY
 from math              import exp, floor, atan2, pi
-from globsim.generic   import ParameterIO, StationListRead, ScaledFileOpen, str_encode, series_interpolate, variables_skip, get_begin_date, create_empty_netcdf, GenericDownload
+from globsim.generic   import ParameterIO, StationListRead, ScaledFileOpen, str_encode, series_interpolate, variables_skip, get_begin_date, create_empty_netcdf, GenericDownload, GenericScale
 from globsim.meteorology import spec_hum_kgkg, LW_downward, pressure_from_elevation
 from fnmatch           import filter
 from scipy.interpolate import interp1d, griddata, RegularGridInterpolator, NearestNDInterpolator, LinearNDInterpolator
@@ -2392,7 +2392,7 @@ class MERRAinterpolate(object):
                                         self.list_name + '_surface.nc'))
 
       
-class MERRAscale(object):
+class MERRAscale(GenericScale):
     """
     Class for MERRA data that has methods for scaling station data to
     better resemble near-surface fluxes.
@@ -2408,17 +2408,7 @@ class MERRAscale(object):
     """
         
     def __init__(self, sfile):
-        # read parameter file
-        self.sfile = sfile
-        par = ParameterIO(self.sfile)
-        self.intpdir = path.join(par.project_directory, 'interpolated')
-        self.scdir = self.makeOutDir(par)
-        self.list_name = par.station_list.split(path.extsep)[0]
-        
-        # read kernels
-        self.kernels = par.kernels
-        if not isinstance(self.kernels, list):
-            self.kernels = [self.kernels]
+        super().__init__(sfile)
             
         # input file names
         self.nc_pl = nc.Dataset(path.join(self.intpdir,
@@ -2467,11 +2457,6 @@ class MERRAscale(object):
                                         units = self.scaled_t_units, 
                                         calendar = self.t_cal) 
 
-        # get the station file
-        self.stations_csv = path.join(par.project_directory,
-                                      'par', par.station_list)
-        #read station points 
-        self.stations = StationListRead(self.stations_csv)  
                                                                                 
         
     def process(self):
@@ -2516,27 +2501,6 @@ class MERRAscale(object):
         self.nc_sf.close()
         self.nc_sa.close()
         #self.nc_sc.close()
-        
-    def getOutNCF(self, par, src, scaleDir = 'scale'):
-        '''make out file name'''
-        
-        timestep = str(par.time_step) + 'h'
-        src = '_'.join(['scaled', src, timestep])
-        src = src + '.nc'
-        fname = path.join(self.scdir, src)
-        
-        return fname
-
-    
-    def makeOutDir(self, par):
-        '''make directory to hold outputs'''
-        
-        dirSC = path.join(par.project_directory, 'scaled')
-        
-        if not (path.isdir(dirSC)):
-            makedirs(dirSC)
-            
-        return dirSC
     
     def PRESS_Pa_pl(self):
         """
